requirements.txt
asgiref==3.8.1
Django==5.1.3
django-modeltranslation==0.19.11
django-utils-six==2.0
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.1
tzdata==2024.2
